long double

+++ includes/fractol.h
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/04 15:20:13 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/10/10 14:23:36 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/10/18 19:03:05 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,13 +32,13 @@ struct					s_vars
 	void				*mlx;
 	void				*win;
 
-	double				scale;
-	double				dx;
-	double				dy;
+	long double				scale;
+	long double				dx;
+	long double				dy;
 
 	int					is_down;
-	double				down_x;
-	double				down_y;
+	long double				down_x;
+	long double				down_y;
 	int					keycode;
 
 	t_data				img;
@@ -49,8 +49,8 @@ struct					s_vars
 	int					color;
 	int					(*model)(int, int, t_vars *);
 	int					is_julia;
-	double				c[2];
-	double				z[WINDOW_WIDTH * WINDOW_HEIGHT * 2];
+	long double				c[2];
+	long double				z[WINDOW_WIDTH * WINDOW_HEIGHT * 2];
 	int					count[WINDOW_WIDTH * WINDOW_HEIGHT];
 };
