+++ src/model.c
@@ -6,17 +6,17 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 12:11:44 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/10/18 17:44:11 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/10/18 19:03:05 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 #include <limits.h>
 
-void	to_z(int w, int h, double z[], t_vars *vars)
+void	to_z(int w, int h, long double z[], t_vars *vars)
 {
-	z[0] = (double)(w - WINDOW_WIDTH / 2) / vars->scale - vars->dx;
-	z[1] = -(double)(h - WINDOW_HEIGHT / 2) / vars->scale - vars->dy;
+	z[0] = (long double)(w - WINDOW_WIDTH / 2) / vars->scale - vars->dx;
+	z[1] = -(long double)(h - WINDOW_HEIGHT / 2) / vars->scale - vars->dy;
 }
 
 void	init_loop(t_vars *vars, int loop)
